Added functionality to Itemizer (#2312)
* Added functionality to Itemizer

Significant changes to Itemizer:
- added auto-stacking of the destination bag when moving partial stacks of stackable items, along with a setting for the functionality, and a manual argument to invoke it.
- changed the previous fix to the recognition of high quality item names (ending in `+1`, `+2`, etc) to include any other non-digit character (i.e `#`)
- changed help text to reflect the new functionality (and fixed capitalization)
- changed all instances of `windower.add_to_chat()` to use `log()` instead, since it's using the `luau` library anyway
- changed version-recognizing logic so that it recognizes changes in major and minor versions automatically without the need for hardcoded values
- Adjusted version check so that it only fires when the user is logged in all the way and thus the chatlog is available to print the notice.
- bumped version number

I am, tentatively, not using the packets library for the stacking functionality, since it adds a good bit of overhead and there can be situations where the addon might be spamming stacking commands, especially when the //put commands are invoked via scripts - as it's common. On the other hand, the addon is already using item resources which is a lot of overhead on its own, so I am not sure if it would be better to use the packets lib as well at that point. I'm open to suggestions.